pyproject.toml
[project]
name = "tunery"
version = "0.4.0"
description = "Build PDF setbooks from sheet music collections"
readme = "README.md"
license = "Apache-2.0" 
authors = [
    { name = "Sasha Ovsankin", email = "sasha@ovsankin.com" }
]
requires-python = ">=3.12,<3.13"
dependencies = [
    "google-cloud-vision>=3.11.0",
    "pikepdf>=9.2.1",
    "pydantic>=2.12.4",
    "pyyaml>=6.0.2",
]

[project.scripts]
tunery = "tunery.main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tunery"]
